+++ functions/run_beyondwords_feed_ea.py
@@ -1,5 +1,5 @@
 from functions.configs import beyondwords_ea_config
-from functions.podcast_feed_generator import update_beyondwords_input_feed
+from functions.feed_updaters import update_beyondwords_input_feed
 
 if __name__ == '__main__':
     update_beyondwords_input_feed(beyondwords_ea_config(), False)
