+++ functions/run_ea_weekly.py
@@ -1,5 +1,5 @@
 from configs import ea_weekly_config
-from podcast_feed_generator import update_podcast_feed
+from feed_updaters import update_podcast_feed
 
 if __name__ == '__main__':
     update_podcast_feed(ea_weekly_config(), False)
