+++ functions/run_lw_all.py
@@ -1,5 +1,5 @@
 from configs import lw_all_config
-from podcast_feed_generator import update_podcast_feed
+from feed_updaters import update_podcast_feed
 
 if __name__ == '__main__':
     update_podcast_feed(lw_all_config(), False)
