+++ tests/test_podcast_feed_generation.py
@@ -1,7 +1,7 @@
 import freezegun
 
 from functions.feed import FeedGeneratorConfig
-from functions.podcast_feed_generator import update_podcast_feed
+from functions.feed_updaters import update_podcast_feed
 from tests.conftest import get_feed_reference_date_str
 
 search_periods = (FeedGeneratorConfig.SearchPeriod.ONE_DAY, FeedGeneratorConfig.SearchPeriod.ONE_WEEK)
